update tinyint -> int

+++ Documentation/Query1EquipeSportive - Copie.sql
@@ -3,17 +3,17 @@ USE Equipe_sportive;
 
 GO
 CREATE TABLE Roles(
-IdRoles TINYINT PRIMARY KEY,
+IdRoles INT PRIMARY KEY,
 Description VARCHAR(20)
 );
 
 CREATE TABLE Etat(
-IdEtat TINYINT PRIMARY KEY,
+IdEtat INT PRIMARY KEY,
 Description VARCHAR(20)
 );
 
 CREATE TABLE TypeEvenement(
-IdTypeEvenement TINYINT PRIMARY KEY,
+IdTypeEvenement INT PRIMARY KEY,
 Description VARCHAR(20)
 );
 
@@ -30,14 +30,14 @@ EstAthlete BIT,
 EstTuteur BIT,
 EstEntraineur BIT,
 EstAdmin Bit,
-FK_Id_Etat TINYINT FOREIGN KEY REFERENCES dbo.Etat (IdEtat)
+FK_Id_Etat INT FOREIGN KEY REFERENCES dbo.Etat (IdEtat)
 );
 
 CREATE TABLE Equipe(
 IdEquipe INT IDENTITY PRIMARY KEY,
 Nom VARCHAR(50),
 Region VARCHAR(30),
-Etat TINYINT FOREIGN KEY REFERENCES dbo.Etat (IdEtat),
+Etat INT FOREIGN KEY REFERENCES dbo.Etat (IdEtat),
 DateCreation DATETIME2,
 DateModification DATETIME2,
 Sport VARCHAR(30),
@@ -51,14 +51,14 @@ Emplacement VARCHAR(200) NOT NULL,
 DateDebut DATETIME2,
 DateFin DATETIME2,
 DateCreation DATETIME2,
-TypeEvenement TINYINT FOREIGN KEY REFERENCES dbo.TypeEvenement (IdTypeEvenement),
-FK_Id_Etat TINYINT FOREIGN KEY REFERENCES dbo.Etat (IdEtat)
+TypeEvenement INT FOREIGN KEY REFERENCES dbo.TypeEvenement (IdTypeEvenement),
+FK_Id_Etat INT FOREIGN KEY REFERENCES dbo.Etat (IdEtat)
 );
 
 CREATE TABLE ConnJoueurEquipe(
 Id INT FOREIGN KEY REFERENCES dbo.Utilisateur (IdUtilisateur),
 FK_Id_Equipe INT FOREIGN KEY REFERENCES dbo.Equipe (IdEquipe),
-FK_Id_Roles TINYINT FOREIGN KEY REFERENCES dbo.Roles (IdRoles),
+FK_Id_Roles INT FOREIGN KEY REFERENCES dbo.Roles (IdRoles),
 );
 
 CREATE TABLE ConnEquipeEvenement(
